Step 2: Creating the skeleton for the information.
Step 2.1 : Creating a section for the main info where it will show a greeting message and the name of the user.

Step 2.2 : Creating another section for the rest of the info using a unorder list to display each one, where each list item have an icon to display the data and a span to wrap the text.

Step 2.3 : Adding an Id and alt atributte to the image.