Step 5: Styling the main element and body background.
Step 5.1 :  Change the background and color of the body.

Step 5.2: Set the background for the main element, text align and padding.

Step 5.3: Set the typography and margin for the intro message and name.

Step 5.4: Style the second section which have the list items in a flex box, modifying the alignment and spacing between them.

Step 5.5: Add classes to the icons to customize the colors and add some margin to the text.