Step 7: Adding the breakpints.
Step 7.1: adding a media query for a min width of 900px.

Step 7.2: Change the layout to be a two columns with the body turn into a flexbox and the main element into a grid.

Step 7.3: Align the profile image to be center at the top.

Step 7.4: Align the text for the information on both sides.

Step 7.5: Add another breakpoint for a min width of 1020px to center the profile picture.