Step 8: Crate the function to handle the API call.
Step 8.1: Add the id to the DOM elements in the html.

Step 8.2: Access to the button on the document and add an Event listener for clicks.

Step 8.3: Create the async function  to handleClickBtn to call the API.

Step 8:4: Access to the data through a call to the API and console log the data to see the data structure.

Step 8.5: Access the elements through the DOM to update the data from the API.

Step 8.6: Do the catch loop to handle any errors with the API call.

Step 8.7:  Add some extra padding in the greeting and name text.